requirements.txt
streamlit
openai
faiss-cpu
sqlalchemy
networkx
sentence-transformers
pickle5
PyPDF2